Updated the draft Unicode BOMs in UTF-8 paper to address feedback from various individuals.
Added an abstract.
Updated the introduction to include wording from the "Unicode Signature"
section of 2.13, "Special Characters".
Reworked the introduction to address different interpretations of the wording
in 2.6 and potential conflicts with 2.13.
Reworked the proposed resolutions to make the options more clear.
Updated the proposed resolution option 1 to remove guidance to also address
section 2.13.
Added an acknowledgements section.